requirements.txt
clarifai
python-dotenv
streamlit